Refactor SeedDerivable trait implementation
