+++ sdk/Cargo.toml
@@ -107,7 +107,7 @@ solana-frozen-abi-macro = { workspace = true, optional = true, features = [
 ] }
 solana-inflation = { workspace = true, features = ["serde"] }
 solana-instruction = { workspace = true }
-solana-keypair = { workspace = true, optional = true }
+solana-keypair = { workspace = true, optional = true, features = ["seed-derivable"] }
 solana-native-token = { workspace = true }
 solana-packet = { workspace = true, features = ["bincode", "serde"] }
 solana-precompile-error = { workspace = true, optional = true }
