+++ sdk/keypair/Cargo.toml
@@ -12,8 +12,11 @@ edition = { workspace = true }
 [dependencies]
 bs58 = { workspace = true, features = ["std"] }
 ed25519-dalek = { workspace = true }
+ed25519-dalek-bip32 = { workspace = true, optional = true }
 rand0-7 = { workspace = true }
+solana-derivation-path = { workspace = true, optional = true }
 solana-pubkey = { workspace = true }
+solana-seed-derivable = { workspace = true, optional = true }
 solana-seed-phrase = { workspace = true }
 solana-signature = { workspace = true }
 solana-signer = { workspace = true }
@@ -25,6 +28,9 @@ wasm-bindgen = { workspace = true }
 serde_json = { workspace = true }
 static_assertions = { workspace = true }
 tiny-bip39 = { workspace = true }
+
+[features]
+seed-derivable = ["dep:solana-derivation-path", "dep:solana-seed-derivable", "dep:ed25519-dalek-bip32"]
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
