+++ sdk/seed-derivable/Cargo.toml
@@ -10,10 +10,7 @@ license = { workspace = true }
 edition = { workspace = true }
 
 [dependencies]
-ed25519-dalek = { workspace = true }
-ed25519-dalek-bip32 = { workspace = true }
 solana-derivation-path = { workspace = true }
-solana-keypair = { workspace = true }
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
