+++ sdk/src/signer/keypair.rs
@@ -1,7 +1,9 @@
 #[deprecated(since = "2.2.0", note = "Use `solana-keypair` crate instead")]
-pub use solana_keypair::*;
-#[deprecated(since = "2.2.0", note = "Use `solana-seed-derivable` crate instead")]
-pub use solana_seed_derivable::keypair_from_seed_and_derivation_path;
+pub use solana_keypair::{
+    keypair_from_seed, keypair_from_seed_phrase_and_passphrase, read_keypair, read_keypair_file,
+    seed_derivable::keypair_from_seed_and_derivation_path, write_keypair, write_keypair_file,
+    Keypair,
+};
 #[deprecated(since = "2.2.0", note = "Use `solana-seed-phrase` crate instead")]
 pub use solana_seed_phrase::generate_seed_from_seed_phrase_and_passphrase;
 #[deprecated(since = "2.2.0", note = "Use `solana-signer` crate instead")]
